requirements.txt
alembic==1.8.0
attrs==21.4.0
blinker==1.4
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
dynaconf==3.1.9
Flask==2.1.2
Flask-Caching==2.0.0
Flask-DebugToolbar==0.13.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-flask==1.2.0
python-dotenv==0.20.0
requests==2.28.0
SQLAlchemy==1.4.39
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
